Issue #5976: Synchronize access to processed injection classes
ReferenceContext.process() will process the majority of injection classes
and will always be called before getProcessedInjectionClasses(), however
processDynamic() may on rare occasions be called concurrently. This
can result in the ConcurrentModificationException or an inconsistent
state in the ReferenceContext.

Synchonizing the getProcessedInjectionClasses() method will ensure that
it cannot run concurrently with processDynamic() (already synchronized).
The ConcurrentModificationException will no longer occur, and any
classes returned as 'processed' will really be fully processed.