Fix race condition when processing timeout
If the running method responds promptly to being interrupted, then
there's a race as to whether the method thread or the timeout callback
thread runs finalizeAttempt first. If the method thread gets there
first, then the result of the method is returned (which may be an
InterruptedException) even though the timeout occurred.

To avoid this, move the stopping of the timeout to processMethodResult
and check timeoutState.isTimedOut() before calling finalizeAttempt.