Make RepeatTests handle exceptions better
RepeatTests needs to cope with the Statement throwing an exception. This
appears to happen if an @BeforeClass or @AfterClass method throws an
exception, which is more likely when the rule is applied to a test class
rather than a suite.

If these exceptions are not caught, then later repeat actions will not
be run.