Check if the Library has been deleted before returning a service
It is possible for a Library to server multiple meta-inf services.
This can cause issues when that library is deleted.  That will
invalidate the factories for all the services the library provides
but we unregister each library one at a time.  If a component
requires one of the services types from a library and that
library provides multiple services of the that same type then
SCR will deactivate the component when we unregister one of the
provided services and attempt to reactivate it with the other
instance of the service before we can unregister it also.

Ultimately this can cause exceptions when trying to create
the invalid service on reactivation of the component.