Updates to the Application classloader to use the parent URL.
The code now will use the parent URL for the shared class cache URL
instead of the class file's URL.  This is what is required for classpath
with directories in them.  It also works for jar files to just pass the
jar with file protocol instead of the jar:file with the entry.  Other
updates to reduce the amount of memory used for different objects in the
AppClassLoader hierarchy.