FT2.0: Catch exceptions completing CompletionStage
Ensure that if an exception is thrown while putting the result into the
CompletionStage, we catch and handle the exception and use it to
complete the CompletionStage.

Exceptions can happen here if there's a problem applying the thread
context.

It's important that we always complete the CompletionStage as the user's
application may be waiting for the result.