Handle async case where app shuts down
If an application shuts down, no more asynchronous operations can be
started for that application. With FT and async, new async operations
can be started at several points for one method call.

This change ensures that in this case, the user gets a sensible error
message, rather than an FFDC and an internal error exception.