Update to initialize Logger before setting application classloader.
When a Logger is created for internal classes, it will use the thread
context classloader to try to load the resource bundle.  This will never
find the resource bundle, but will end up looking up 4 resources in
order to find that out.  resourceBundle.class, resourceBundle_en.class,
resourceBundle_en_US.class, and resourceBundle.properties.  Initializing
the classes before setting the application classloader will allow Logger
to skip the classloading of these resources from the application
classloader and only look in the caller classloader.