Issue #7434-Fix NPE with Stateful access-timeout
EJB 2.x stateful beans (with a home) have an extra method added to the
list of method attributes to handle the remove method on the home. metadata
for the method is set, but not the method reflect object, which is
used by MethodAttribUtils; so an NPE occurs processing <access-timeout>
for such a scenario.

A null check has been added to skip further checking; access-timeout would
never apply to the home method.