Create a collection of built-in classes/interfaces used as Providers.
Presently there is a lot of processing that is done to determine if a
Provider has a @Context annotation on a field and/or method.  This
affects start up time.  This change creates a collection of CXF and
JAX-RS classes and interfaces so that we know if we need to do the
expensive reflection processing to find @Context annotations.  If the
known classes do not have method and/or field @Context annotations, we
can skip one or both of the field or method processing, or not process
the Provider class at all.