Make Config bean ApplicationScoped
Rather than finding the classloader when the Config bean is created, do
it on each call to the bean. This allows us to make the config bean a
singleton.

This shouldn't greatly impact performance since each call to a
RequestScoped bean will need to check the request scope of the current
thread anyway.

Making the Config bean ApplicationScoped allows it to be used at times
when the RequestScope isn't active (e.g. during extension lifecycle
events).