Allow extension BDAs to see other extension BDAs
Allow BDAs for runtime extensions to to see other runtime extension BDAs
if they are able to see application BDAs.

This allows runtime extensions which treat beans as services to see the
beans they expect to (e.g. the mp reactive messaging extension to see
the Config bean from mpConfig).