Create a new ApplicationInfo object if the Application was replaced.
if CDI or EJB replace the application object with a different object we
need to create a new ApplicationInfo object like what was done
previously always.  It will only create a new ApplicationInfo object
though if the replace happens instead of how it was doing it
unconditionally before which is what caused the perf problem.