Minor updates to not load as many classes.
- Update IBMMultipartProvider to not have reference to the classes and
only reference the class names.  The class collection isn't needed
because we already check the class names so it is just extra work to
check the class collection as well.  Also we don't need to load the
classes if they aren't used.
- Update WSThreadLocal to remove the logic related to ThreadPool.Worker
since we don't use the ThreadPool implementation, we can just not have
that logic.
- Update references to com.ibm.ws.config.  container.service didn't
directly depend on it.  It appears to only have been there for other
components to depend on getting stuff from com.ibm.ws.config through the
container.service.