WEB:  Clean up references to applications in web container code
- Update logic that sets context on a ThreadLocal so that we unset it in
a finally.  If an exception happens we end up leaving context on a
ThreadLocal and that context has a reference to the application
ClassLoader.
- When doing async start of a web application make sure that if there is
an error starting that we stop the module just like we do if we are not
doing async start of web applications.
- If we fail to complete initialization allow the destroy to happen
still so that state is cleaned up.  Otherwise we leave references to
application classloaders and other things.