Allow registration update of public OAuth clients to use asterisk for client secret
It looks like the OAuth client registration code has a bug in which a public client whose token endpoint auth method is set to `"none"` cannot have `"*"` used as the client secret in a PUT request to the registration endpoint. A value of `"*"` is supposed to indicate that the client secret should remain unchanged, but an error is returned in the case where the client is public and has its token endpoint auth method set to `"none"`.