Refactor Kafka test suites
Different tests will need Kafka brokers started in different
configurations.

Split the tests into sub-suites based on which Kafka broker they need to
run with. This avoids having the broker running for longer than it needs
to and makes it easier to avoid paying the time cost of starting the
broker in every configuration during development or in LITE mode.