Set com.ibm.ws.app.manager to CONTAINER_EARLY start-phase
The component com.ibm.ws.app.manager.internal.ApplicationConfigurator
has a load of synchronization in its methods that are called by
SCR whild holding the SCR implementation is holding the component
state change lock.  This can lead to certain deadlock which is
eventually broken out of, but results in the OSGi service
registry eventually returning null when other components are
trying to get the services provided by the ApplicationConfigurator
component.  To reduce this posibility the start-phase of the
com.ibm.ws.app.manager is being moved up so that this component
is fully activated (it is immediate) before the CONTAINER
start-phase is entered.