Add FT Metrics 2.0 guard feature
This guard feature prevents the FT-Metrics-2.0 feature from being
included automatically in the release package because both mpFT-2.0 and
mpMetrics-2.0 are GA. This causes validation errors because the
autofeature itself is not GA and so should not be in the release
package.

Add this extra guard feature so that it won't get included in the
release package but we can still include the guard feature so we can
develop and write tests with it until it is ready to be released.