Rename the FT 2.0 + Metrics 1.1 test repeat config
Rename the test config FAULTTOLERACE20 to FT20_METRICS11

The set of features in this config remains the same

The original name was written before FT 2.0 was included in a
microprofile release, and so it was the old MP release, plus the new
version of Fault Tolerance.

Now that Fault Tolerance has been included in an MP release, the name
makes little sense and has confused too many people already.