Avoid adding metrics to all Fault Tolerance tests
Change the repeat tests configurations to set forceAddFeatures to false
so that it doesn't add mpMetrics-x.x to tests which don't already have
metrics in the server.xml

This also requires adding the guard feature to all the server.xmls which
use mpMetrics-2.0. This feature does nothing except for enabling the
autofeature integrating FT and metrics-2.0.

Additionally update some of the server.xmls so that they have a
consistent set of MP3.0 features. This doesn't actually have any effect
because the feature list gets overwritten by the repeat tests action,
but it avoids having a set of features we don't actually test in the
server.xml.