Use HttpUtils to make actual user API call
Status: Full authentication and authorization flow should work; should be able to successfully reach the protected resource.

The user info response is being manually edited right now to satisfy the `userNameAttribute` requirement of having a specific attribute in the response JSON that's used as the username. For now, I have that hard coded to be the user `ayoho-edited-username`.