Issue #9479: Support RMI Remote throwing Exception
JITDeploy has been updated to allow an RMI Remote method to declare throwing Exception
rather than requiring it throw RemoteException.  Spec indicates it must be RemoteException
or a parent of RemoteException.

Generatioin of EJBWrapper also updated to add catch block for RemoteException so that
it is not treated as a checked exception, like other Exception subclasses....
except when both Exception and RemoteExctption have been declared. Both Exception
and RemoteException together were allowed previously, so don't want to change the
existing behavior.