Improve equinox trace enablement
- Make TrOSGiLogForwarder a LogFilter as well and add it when adding the
LogListener.  Without a filter all trace is considered enabled and does
extra processing when trace isn't actually enabled for the logger.
- Update the map of Liberty log levels to equinox log levels to only do
debug if it is finest/debug and not set it to debug for fine/event and
finer/entryexit because it won't log debug anyway when it is that
liberty level.