Allow only running one repeat of a FAT bucket
There was already a property to only run a single repeat, but it wasn't
possible to set it from the command line when running locally.

Not all properties get passed through to the test process, but
properties which start "fat.test" do, so change the property name to
"fat.test.repeat.only".