Avoid closing entityStream in JsonBProvider
This is required by the spec (see MBW's javadoc) that we must not close
the entity stream in the writeTo method. The toJson(obj, stream) method
will always close the stream, so this cannot be used.

This commit includes a test case that verifies that the client response
filter is only called once per response.  The prior behavior invoked
the filter twice - once as a result of the toJson(obj, stream) method
closing the stream and again when the response was actually processed.