Use proper TCCL during CDI startup
Previously, we were setting the TCCL to the app classloader. This allows
access to application classes, but does not allow access to internal
classes exposed through the TCCL. This is important for certain
specifications which rely on service loading their provider through the
TCCL.

This is particularly a problem for reactive messaging because
user-provided code can be called during the application startup and can
reasonably expect to be able to access these features.