Issue #10453: Auto-link EJB ref in java global with ejb-link
Relax the checking in EJBProcessor to allow ejb references
in java:global to use auto-link when the application name
is known and ejb-link has been specified (i.e. ejb-link or beanName).

This will allow EJB references in java:global to auto-link to
an EJB specified by ejb-link that exists in the same application.

If the EJB is in a different application, a binding will still be
required.

Also removed a binding from a test that was previously added to
demonstrate the current behavior. The test will now properly link
to the EJB without the need for a binding.