Kafka connector redesign
Redesign the Kafka connector slightly to give a stronger pairing between
a message and the partition tracker associated with it. This prevents
confusion when a partition is unassigned and then reassigned.

Avoid emitting messages if the partition those messages came from has
since been unassigned (and therefore an attempt to commit the partition
offset would fail).