Issue 10277: Autonomical algorithm for calculating pollInterval.
We store two values in the partition table.
(1) last time we pulled
(2) time when the last poll slot reserved will trigger

We will assign the current poll to use value (2) plus a global poll value to determine when it will poll for tasks next.
Value (1) is used to detect if any poll intervals have been missed.