Fix timestamp bug in ConsumerRecordTest
The ConsumerRecordTest manually set the timestamp to 100 (i.e. long ago
in the past). After an indeterminate period, the server would notice
that this was before the retention period and delete it, causing the
test to intermittantly fail.

Replace this with a more realistic timestamp, set the same data into a
header and confirm that they match in the received message.