Add an arquillian extension and impl of LRARecoveryService
The arquillian extension means that LRARecoveryServiceImpl gets added to
all the war files that arquillian creates.

The LRARecoveryServiceImpl is required for the tests to run, as it allows
the tests to poke the internals of the LRA co-ordinator. At this point
the impl is a noddy impl which just implements the interface and doesn't
communicate with the co-ordinator at all.

If the co-ordinator becomes a liberty feature, then this arquillian
extension will need to be dropped, and the LRARecoveryServiceImpl will
have to be included in a 'test only feature', so that it can talk to
the co-ordinator, and still be accessible to the test code