Issue #11089: Fix non-existing timer behavior
When an EJB timer is created with no expirations, it is considered to
not exist, and any method called on it should throw a NoSuchObjectLocalException.

Persistent EJB timers behave properly, as well as the timer.getNextTimeout()
method for non-persistent timers.

Fix the other timer methods by moving the code that checks the state and
throws the exception from the getNextTimeout() method to a common code
path for all timer methods.