New sec.authorization.util project for role-based auth
This new project is intended to be used by components that perform
authorization checks using `@DenyAll`, `@PermitAll`, and `@RolesAllowed`
annotations. It handles the busy work of checking whether the target
class contains these annotations or whether the target method contains
these annotations (in which case the method overrides the class
annotation).