Issue #11572: Fix Calendar EJB timer with no expiration
A calendar EJB timer with no expiration may be returned by getTimers()
if getTimers() is called in the same transaction that creates the timer.
A timer with no expiration is not active and should not be returned by
getTimers().

Updated TimerNpImpl to expose the expiration (i.e 0 means none) and
then updated ContainerTx to not queue a timer for starting if it in
fact has no expirations and will thus never expire. Since the timer
is not queued to start on tx commit, the getTimers() will not find it.