Adjust EJB Calendar Timer Create Time
The first expiration of a calendar based EJB timer must be after the
time of timer creation and after the specified start date/time.

Adjust this calculation to account for any time delay between the time
the API is called to create the timer and when the timer is actually
scheduled to run.