Switch from JCache API to Infinispan API via reflection for cache creation.
When Liberty connects to Infinispan servers through the JCache API and hotrod client, the infinispan.xml configuration is getting lost. To work around this we are using the Infinispan API directly to ensure caches are replicated throughout an Infinispan cluster. This work is done via reflection to prevent dependencies on Infinispan code.