Issue #11919: Unblock EJB remtoe app start on shutdown
When EJB remote apps start, the start process waits for the ORB to
start before proceeding. This wait should be discontinued if the
server is then stopped.

Changed to discontinue the wait when EJB container is notified the
server is stopping.

Also, once the ORB starts, fixed an order problem in the container
that will bind all remote interfaces. The latch used for the wait
needs to be cleared first, otherwise the bind process will wait
for 30 seconds, despite the ORB already having started.