Issue #12102: Avoid Token error on server stop
Despite the test waiting for messages indicating security has started,
occasionally it hasn't fully started and this results in the error:

CWWKS4000E - The requested TokenService instance of type Ltpa2 could not be found.

This is caused by the InitTxRecoveryLogBean that is an @Startup bean
that runs as soon as the server start.

To avoid this... moving the install of the @Startup bean until after
the messages appear in the server indicating security has started.

Also, updated cases where test is waiting for a config update to also
wait for the InitTxRecoveryLogBean to start, to avoiding stopping server
while it is still attempting to start.