Improve resource injection trace
Since supporting Object Factories are injected into the injection
engine, it can be difficult to tell which component is injecting
an object facotry where there is a failure.

Include the name of the object facotory in trace and the exception
message when an error occurs in registration.

Also improve trace when processing class annotations.