Issue 12778: Fix JWK cache logic
A change delivered in https://github.com/OpenLiberty/open-liberty/pull/9597 can cause the logic to return `null` earlier than expected when trying to find a cached JWK. This can happen if Liberty falls back to using the key text to look up the key in the cache and fails to find a match. The logic is being updated to go back to falling back to checking the cache that's keyed off the JWK location.

Resolves #12778