Issue #13066: Add deserialization provider for CDI 3.0
Added DeserializationClassProvider for CDI 3.0 so that objects
injected into EJB stateful beans, or interceptors provided by
CDI may be serialized and deserialized as part of stateful
bean passivation.

The EJB Container does not need to be aware of the specific
CDI classes, nor do that need to be exported from the CDI bundle.