Issue #13249: Change EJB lifecycle tx complete to tx end
The "Unspecified" context are processed in EJB Container only for lifecycle
callbacks that cannot otherwise have a transaction attribute configured.
The EJB Container is managing the "boundary", by calling begin/complete
around the lifecycle interceptor. What is important is to consider who
is responsible for the commit or rollback of the work. Since the code is
application code, it seems that the resolver should default to application
and only become "container" if so configured by the applications.

The correct thing to do here is to switch from LTC.complete() to LTC.end().
LTC.end() checks for ContainerAtBoundary, and either calls LTC.complete(),
or does the appropriate resolver action determined by the application.