Revisions to address FAT test failure.
Added javaPermission elements to resolve Java 2 security errors.

Revised lookup of ports to abort if there is a failure making it clearer
when it is the lookup that fails first.  Added explicit reporting of
looked-up values to aid diagnosis should similar problems occur again.

Changed exception stack reporting to go to System.out in an effort to
ensure failures are not missed in the testing framework logs. (Should be
unnecessary but test failure seems to be missing an exception stack in
order to explain what was seen.)

Removed all the automatically generated TODO comments.