Make FT 3.0 use Metrics 3.0 instead of 2.3
MP Fault Tolerance 3.0 will release alongside MP Metrics 3.0, so we need
them to work together.

This PR changes the metrics integration to use 3.0 instead of 2.3.
Breaking changes in Metrics 3.0 mean that the same jar cannot support
both versions of MP Metrics. (The source code will actually compile and
run against either, but the built binary will only be compatible with
the version it was compiled against.)

An additional autofeature and bundle could be added later to support
Metrics 2.3 as well if required.