Check for alternate feature names and log new related error message.
Additional meta-data can be added to a feature definition for commonly known
alternate names. If a configured feature is not found during a feature
configuration update and it matches a known alternate name, then an error
message will be issued with the real name of the feature as a hint.

closes https://github.com/OpenLiberty/open-liberty/issues/13393