Disable all features on conflict
A new feature header is introduced

WLP-DisableAllFeatures-OnConflict: true|false

If not specified the default is false.

When the configured feature set has a singleton conflict
and the resolved feature set or the conflicting candidates
has a feature with a value of true for this new heade then
all features will be disabled.  In other words,
no features will be installed for the server/client
instance.