Increase timeout for receiving bundle events
Bundle events are asynchronous for BundleListeners
The testcode and the runtime code that tracks bundle origin
both use the asynch BundleListener to track the events.

This could be done with a SynchronousBundleListener
but there is no need for the clean up of the tracked
bundles to be immediate and it risks slowing down
the worker thread for no reason (and potential
for deadlock).

The timeout for tracking the events in the test
has been increased to 30 seconds.  Also the
test BundleActivator has be modified to do
the actual test work in a background thread
instead of synchronously from BundleActivator.start