Issue #14282: Ignore CWWKG0014E in EJB FAT
CWWKG0014E occurs when the FAT updates server.xml while the server
is running and the kernel begins to process the file before it
is fully written. The kernel will retry and the test will run
fine, so the error is just an odd timing problem and may be ignroed.

Updating all EJB Container FAT that does frequent server config
changes while the server is running.