Issue #14475: Fix AmbiguousEJBReferenceException in App Client
Expose this Liberty API class in the app client just like it is
exposed in the app server. This will allow an app client application
class to catch this exception if it is thrown by the server process
becauase the lookup or injection of an EJB is ambiguous (i.e. multiple
EJB implementations match the criteria).

Also fix a similar problem for a Jakarta EE server process.