Update to not format method parameters when audit isn't enabled
- Whether audit is enabled or not we were formatting the method
parameters.  Moved to do the formatting where it is actually used.
- Additionally the maps used for the passing of ejb invocation data were
class scoped which is wrong and should have been method scoped.  This
could result into memory leaks as was fixed in PR #7987 and hang
conditions when multiple threads are accessing the HashMap concurrently
since it is not thread safe.