Ensure manifest file for Liberty packaged server archive
The temporary manifest File for the Liberty server archive is not closed
after writing it.  It is still left open when trying to read it to add
it to the final archive.  It is possible this is causing some corruption
of the manifest in the final archive.

Also adding trace to the SpringBootThinUtil to help debug the inputs
during the Spring Boot FAT that use the Liberty uber JAR.