Issue #14931: Fix EJB Handle in App Client
Jakarta EE 9 App Client is currently failing if an EJB Handle
is returned to the client with a class not found for
com.ibm.ws.ejb.portable.HandleImpl, the liberty implementation.

Fixed by exposing as internal API like it was for Java EE Clients,
but did that in ejbRemoteClient feature instead of Jakarta EE Client
as was done for older levels.

Also, added FAT tests that covered this scenario, as well as other
scenarios that test ejb-ref resolution in the application client
using ejb-link/beanName and auto-link.