Disable osgi.checkConfiguration by default
When enabled this setting cause Equinox
to check the timestamp of every bundle JAR
installed and it uninstall anything that
has changed since last start.

The framework then causes a refresh of the
uninstalled bundles.  In some cases this
will cause other bundles to re-resolve
which ends up resolving before the
region digraph has been registered with
the framework.  This means the bundles
will be resolved without the issolation
which the Feature Manager establishes.

It appears in some cloud environments and
container images false timestamp reporting
is happening which incorrectly causes some
kernel bundles to be tossed.  To avoid
this the osgi.checkConfiguration option
is disabled.