mpOpenAPI-2.0: Fully scan module directories
When a web module corresponds to a directory on disk, scan the whole
directory, not just the WEB-INF/classes directory, so we don't miss the
contents of WEB-INF/lib.

This makes the behaviour consistent whether the module is an archive or
a directory. However, in both cases we're potentially scanning more than
we should.