OpenAPI-2.0: write cache entries asynchronously
Part of the startup process is to write the generated OpenAPI model to
the cache so that it can be used next time the application starts.

However, since the cache entry isn't needed for the current run of the
application, we can write it in the background to allow the app to start
faster.