Issue #15293: Retry EJB stateful timeout FAT
Test is verifying that the EJB stateful time can be set to 15 seconds
and the bean won't timeout until after that; however, if the VM running
the test has a slight pause while in the middle of the test, then it
may appear the bean timed out before 15 seconds.

If the difference in time from when the bean was last accessed and the current time
is > 15 seconds, then repeat the test.