Adds a service loader reference to CDIRuntimeImpl (which implements CDIProvider)
This is a workaround for eclipse-ee4j/cdi#461. Because
of that bug CDI.java will clear the CDIProvider set in AbstractCDIRuntime.start()
if CDIProvider.getCDI() ever returns null and later attempt to reacquire it
through the service loader.

CDIProvider.getCDI() will return null if ever called from the context of a non
CDI app