Wait for system WABs before registering ServerStarted service
A new service type called ServerReadyStatus is introduced
which the feature manager will call before registering
ServerStarted.  Implementations of ServerReadyStatus
have the check method called to allow them to block
while waiting for the server to be ready before
proceeding to register ServerStarted.

This allows the WABInstaller to check on the
status of the system WABs to make sure they
have completed their provisioning to the web
container