Correct handling when auto-features introduce a conflict
When an auto-feature introduces a new decision that
introduces conflict the resolver will backtrack the
decision available which the auto-feature enables
with tolerates. The issue is the snapshot of the
previously locked in preResolved features is
holding a Set that gets modified while resolving
the conflict the auto-feature introduced. This ultimately
leads to bad set of resolved features that includes
invalid selections and at times multiple versions
of the same singleton feature.

The solution is two parts.  One is to make sure
we clear previous results when setting a new
resolution set. The other part is to make sure
a copy of the resolved feature set is made when
processing newly selected auto features so we
can keep a constant set of preResolved features
as the resolver tries the various permutations
looking for the best solution.