Move checkpoint "spi" into kernel boot to allow hook impl in kernel
Without the SnapshotHook in the kernel no kernel components
can take part in the prepare/restore phase of checkpoint.
Only moving the interfaces for the "spi" to kernel, not
the implementation.  Also note that this is not real
SPI until we have a feature declare it as SPI. For
now this is a Liberty internal package