Set the TCCL correctly during mpOpenAPI-2.0 setup
Ensuring that app classes can be loaded via the TCCL allows filters to
make use of MP Config and find all of the correct config sources.

We must use the ClassLoadingService to create the TCCL so that
ServiceLoader works. In particular, Smallrye OpenAPI uses ServiceLoader
to find its JAX-RS extension.

ApplicationProcessor and ApplicationRegistry were converted to
components to make it easier to inject the ClassLoadingService.