Issue 15023: Fix WASReqURLOidc encoding problem
Fixes #15023

Instead of encoding the query string in the original request URL, we'll leave it as-is. That should ensure we ultimately use the "real" value of the query string. During the creation of the WASReqURLOidc cookie, the cookie value is encoded anyway to mitigate against malformed and potentially malicious cookie values. The `OidcRedirectServlet` code is also updated to use slightly different code to read the cookie so that the respective decoding is done.

15023: working